requirements.txt
Glymur==0.12.9.post1
imageio==2.34.0
lxml==5.1.0
memory-profiler==0.61.0
numpy==1.26.4
objgraph==3.6.1
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
psutil==5.9.8
tqdm==4.66.2
